A nice thing and a bad thing.
Good: I've moved the tests under a tests subdir, a.l.a. kdelibs.
To compile them:
$ cd tests
$ unsermake check
In this way, less code is compiled and installed as default

Bad: I've commented with #if 0 a pair of constructors that don't compile
here, and I don't know how to fix, sorry :(

CCMAIL: kstars@30doradus.org


svn path=/trunk/KDE/kdeedu/libkdeedu/; revision=456393