fix selection of multiple atoms (Carsten: I apologize for the lack of documentation of my rendering code. Currently you have to call m_sphere.select() before you can draw a sphere. I could change that now to make it more easy to use, but it's not the right time to do it because a lot of things are going to change when we'll implement rendering different atoms with different sizes).
also begin implementation of text renderer for the GL widget. The 
problem with QGLWidget::renderText is that it only works for ASCII text 
(possibly 8-bit-extended, though I don't know how to set the character 
encoding, but in any case no unicode) . Thus no 
safe i18n is possible. Even if it's not necessary to display the message 
"please load molecule", I thought it'd be nice to have the possibility 
to display whatever text we want in the QGLWidget. But if you think it's 
futile, I'll drop it.

Benoit


M    kalzium/src/kalziumglwidget.h
M    kalzium/src/kalziumglhelperclasses.h
M    kalzium/src/kalziumglwidget.cpp

svn path=/trunk/KDE/kdeedu/kalzium/src/kalziumglwidget.h; revision=555816