Change some things in the way OpenGL selection is done
It now happens on button press, instead of happening on button release. 
The clicked atom is highlighted in a separate color (currently some 
shade of purple, tell me what you think) even if it is not selected. The 
point of that is that this clicked atom is going to play a very 
important role in the mouse navigation system I plan to write. Any 
rotation/move will be done with respect to it. So I thought that this 
special role of this atom should correspond to a visual highlighting.

M    kalzium/src/kalziumglwidget.h
M    kalzium/src/kalziumglwidget.cpp

svn path=/trunk/KDE/kdeedu/kalzium/src/kalziumglwidget.h; revision=566143