Implement new fields: pregrade and interval
This patch implements two new fields in a kvtml file:
 - pregrade, which is equal to grade but used in the interval 0-1 days
   as opposed to grade which is used for intervals 1 day and up.
 - interval, which is used to indicate how long since a word was last
   trained it is due for training again.

This is prerequisite to the implementation in Parley of the initial
training. The reason for implementing these two fields instead of
changing old ones is that it will still be backward compatible. I
think this is important, especially in the light of the ktouch issue
that we saw on the mailing list a few days ago.

REVIEW: 117730