Create a unit test framework to allow API and non API testing
This framework allows some of the unit tests to be linked to the
shared libraries of the kvocdoc API and others to be linked to all
of the internal objects.  The idea is that more simpler unit tests
will be a more robust approach to testing.  It is also easier to add
small tests around individual objects that do everything at once.

(cherry picked from commit cb91ebfb16e5806382c8821054420c6db47d03d6)