Fix error when destroying an default constructed KEduVocDocument
When a KEduVocDocument which was constructed by the default constructor
becomes destructed, the error message "QFile::remove: Empty or null
file name" is given on stderr.

Since the file name of the KAutoSaveFile was not set in the default
constructor, the KAutoSaveFile::remove() method in ~KEduVocDocument()
have given this error message. Removed this method call and the
following KAutoSaveFile::releaseLock(), because they are called in the
KAutoSaveFile destructor as well but with check for file name.

Added unit test emptydocumenttest to proof the bugfix.

REVIEW: 127717