Fix a bug in SharedKvtmlFiles.
When kanagram or khangman hold a document open, then their
settings dialog scans for files it's unable to get the lock on the one
in use, so that file shows with no Title or Description.
Since we are just reading the files ignore locks and read the data
when iterating the list of shared kvtml files.

Also removed an unused duplicate variable.